Update: separate auth controller from user controller
